replace time_constants.h with nidas/linux/types.h
fix code which broke with the LL contants, especially std::max() calls
which were comparing an int to a long long.

add a test that expressions with an int do not overflow.

move some implementation out of headers so contants do not need to be
included.

+++ src/nidas/dynld/SampleOutputStream.cc
@@ -46,7 +46,7 @@ SampleOutputStream::SampleOutputStream():
     _maxUsecs(0),_lastFlushTT(0)
 {
     _maxUsecs = (int)(getLatency() * USECS_PER_SEC);
-    _maxUsecs = std::max(_maxUsecs,USECS_PER_SEC / 50);
+    _maxUsecs = std::max(_maxUsecs, int(USECS_PER_SEC / 50));
 }
 
 SampleOutputStream::SampleOutputStream(IOChannel* i, SampleConnectionRequester* rqstr):
@@ -54,7 +54,7 @@ SampleOutputStream::SampleOutputStream(IOChannel* i, SampleConnectionRequester* 
     _maxUsecs(0),_lastFlushTT(0)
 {
     _maxUsecs = (int)(getLatency() * USECS_PER_SEC);
-    _maxUsecs = std::max(_maxUsecs,USECS_PER_SEC / 50);
+    _maxUsecs = std::max(_maxUsecs, int(USECS_PER_SEC / 50));
     _iostream = new IOStream(*getIOChannel(),getIOChannel()->getBufferSize());
     setName("SampleOutputStream: " + getIOChannel()->getName());
 }
@@ -68,7 +68,7 @@ SampleOutputStream::SampleOutputStream(SampleOutputStream& x,IOChannel* ioc):
     _maxUsecs(0),_lastFlushTT(0)
 {
     _maxUsecs = (int)(getLatency() * USECS_PER_SEC);
-    _maxUsecs = std::max(_maxUsecs,USECS_PER_SEC / 50);
+    _maxUsecs = std::max(_maxUsecs, int(USECS_PER_SEC / 50));
     _iostream = new IOStream(*getIOChannel(),getIOChannel()->getBufferSize());
     setName("SampleOutputStream: " + getIOChannel()->getName());
 }
