+++ src/nidas/linux/diamond/dmd_mmat.c
@@ -1926,8 +1926,8 @@ static void dmmat_a2d_bottom_half(void* work)
                                 int tdiff = tt0 - tt0orig;
                                 KLOG_WARNING("%s: time tags shifted later by %d.%04d seconds\n",
                                         getA2DDeviceName(a2d),
-                                        tdiff / TMSECS_PER_SEC,
-                                        (int) abs(tdiff) % TMSECS_PER_SEC);
+                                        (int) (tdiff / TMSECS_PER_SEC),
+                                        (int) (abs(tdiff) % TMSECS_PER_SEC));
                         }
                 }
 
@@ -1999,8 +1999,8 @@ static void dmmat_a2d_bottom_half_fast(void* work)
                                 int tdiff = tt0 - tt0orig;
                                 KLOG_WARNING("%s: time tags shifted later by %d.%04d seconds\n",
                                         getA2DDeviceName(a2d),
-                                        tdiff / TMSECS_PER_SEC,
-                                        (int) abs(tdiff) % TMSECS_PER_SEC);
+                                        (int) (tdiff / TMSECS_PER_SEC),
+                                        (int) (abs(tdiff) % TMSECS_PER_SEC));
                         }
                 }
